requirements.txt
docker_py==1.10.6
python_magic==0.4.15
